TextClipPadding replaced with UnclippedTextRenderer
